JA-style API Method for getting the names of all furniture in a given content pack.
Introduces an API class with a single method to return, given the unique id of a content pack, a list of all furniture names defined by that content pack.

The intent of providing this API method is to facilitate integration with ChroniclerCherry's ShopTileFramework, allowing custom furniture to be sold in shops other than Robin's carpentry shop.